issue-276 spec for plain Datasource on destroy
